+++ Core/Src/main.c
@@ -143,7 +143,7 @@ void HAL_GPIO_EXTI_Callback ( uint16_t GPIO_Pin )
 }
 
 //========================================================================
-void LoopProcMain( int nTick )
+void LoopProcMain( uint16_t nTick )
 //========================================================================
 {
 	//	Watchdog Reload
